Move code from check_list_item to the new DueDate component
