+++ webapp/src/components/checklist_item.tsx
@@ -15,7 +15,6 @@ import {useDispatch, useSelector} from 'react-redux';
 import {components, ControlProps} from 'react-select';
 import styled, {css} from 'styled-components';
 import {DraggableProvided} from 'react-beautiful-dnd';
-import {DateTime} from 'luxon';
 
 import {handleFormattedTextClick} from 'src/browser_routing';
 import {
@@ -42,8 +41,9 @@ import MarkdownTextbox from 'src/components/markdown_textbox';
 import CommandInput from './command_input';
 import GenericModal from './widgets/generic_modal';
 import {BaseInput} from './assets/inputs';
-import DateTimeSelector, {DateTimeOption, optionFromMillis} from './datetime_selector';
+import {DateTimeOption} from './datetime_selector';
 import {Mode} from './datetime_input';
+import DueDate from './checklist_item/duedate';
 interface ChecklistItemDetailsProps {
     checklistItem: ChecklistItem;
     checklistNum: number;
@@ -361,17 +361,6 @@ const ControlComponent = (ownProps: ControlProps<ProfileOption, boolean>) => (
     </div>
 );
 
-const ControlComponentDueDate = (ownProps: ControlProps<DateTimeOption, boolean>) => (
-    <div>
-        <components.Control {...ownProps}/>
-        {ownProps.selectProps.showCustomReset && (
-            <ControlComponentAnchor onClick={ownProps.selectProps.onCustomReset}>
-                <FormattedMessage defaultMessage='No due date'/>
-            </ControlComponentAnchor>
-        )}
-    </div>
-);
-
 const portal: HTMLElement = document.createElement('div');
 document.body.appendChild(portal);
 
@@ -444,39 +433,6 @@ export const ChecklistItemDetails = (props: ChecklistItemDetailsProps): React.Re
         }
     };
 
-    const makeSuggestedDateTimeOptions = (mode: Mode.DateTimeValue | Mode.DurationValue) => {
-        let dateTime = DateTime.now();
-        dateTime = dateTime.endOf('day');
-
-        const list: DateTimeOption[] = [];
-        list.push(
-            {
-                ...optionFromMillis(dateTime.toMillis(), mode),
-                label: formatMessage({defaultMessage: 'Today'}),
-                labelRHS: (<LabelRight>{dateTime.weekdayShort}</LabelRight>),
-            }
-        );
-
-        dateTime = dateTime.plus({days: 1});
-        list.push(
-            {
-                ...optionFromMillis(dateTime.toMillis(), mode),
-                label: formatMessage({defaultMessage: 'Tomorrow'}),
-                labelRHS: (<LabelRight>{dateTime.weekdayShort}</LabelRight>),
-            }
-        );
-
-        dateTime = dateTime.plus({days: 6});
-        list.push(
-            {
-                ...optionFromMillis(dateTime.toMillis(), mode),
-                label: formatMessage({defaultMessage: 'Next week'}),
-                labelRHS: (<LabelRight>{dateTime.toLocaleString({weekday: 'short', day: '2-digit', month: 'short'})}</LabelRight>),
-            }
-        );
-        return list;
-    };
-
     const [profileSelectorToggle, setProfileSelectorToggle] = useState(false);
     const assignee_id = props.checklistItem.assignee_id; // to make typescript happy
 
@@ -489,12 +445,6 @@ export const ChecklistItemDetails = (props: ChecklistItemDetailsProps): React.Re
     };
 
     const toggleDescription = () => setShowDescription(!showDescription);
-
-    const [dateTimeSelectorToggle, setDateTimeSelectorToggle] = useState(false);
-    const resetDueDate = () => {
-        onDueDateChange();
-        setDateTimeSelectorToggle(!dateTimeSelectorToggle);
-    };
 
     const content = (
         <>
@@ -547,25 +497,10 @@ export const ChecklistItemDetails = (props: ChecklistItemDetailsProps): React.Re
                                     controlledOpenToggle={profileSelectorToggle}
                                     showOnRight={true}
                                 />
-                                <DateTimeSelector
+                                <DueDate
                                     date={props.checklistItem.due_date}
                                     mode={Mode.DateTimeValue}
-                                    onlyPlaceholder={true}
-                                    placeholder={
-                                        <HoverMenuButton
-                                            title={formatMessage({defaultMessage: 'Add due date'})}
-                                            className={'icon-calendar-outline icon-16 btn-icon'}
-                                        />
-                                    }
-                                    suggestedOptions={makeSuggestedDateTimeOptions(Mode.DateTimeValue)}
                                     onSelectedChange={onDueDateChange}
-                                    customControl={ControlComponentDueDate}
-                                    customControlProps={{
-                                        showCustomReset: Boolean(props.checklistItem.due_date),
-                                        onCustomReset: resetDueDate,
-                                    }}
-                                    controlledOpenToggle={dateTimeSelectorToggle}
-                                    showOnRight={true}
                                 />
                                 <HoverMenuButton
                                     title={formatMessage({defaultMessage: 'Edit'})}
@@ -780,10 +715,3 @@ const ChecklistItemEditModal = (props: ChecklistItemEditModalProps) => {
         </GenericModal>
     );
 };
-
-const LabelRight = styled.div`
-    font-weight: 400;
-    font-size: 12px;
-    line-height: 16px;
-    color: rgba(var(--center-channel-color-rgb), 0.56);
-`;
